Harmonize helper functions
